refactor(server/utils): rename zodErrors to errors
