+++ sandbox-server/src/services/gitServiceFactory.ts
@@ -1,7 +1,7 @@
 import simpleGit from "simple-git";
 
 import { GitService, type IGitService } from "~/services/gitService";
-import { unwrapErrorMessage } from "~/utils/zodErrors";
+import { unwrapErrorMessage } from "~/utils/errors";
 
 export async function createGitService(
   repoPath: string,
