+++ sandbox-server/src/utils/errors.ts
